Create WebDev-HTML-CSS
This is a starter page for creating a new html file.  A short amount of css styling is included in the <head> section.  Order of items are important to follow.  Line indent style recommended for HTML/CSS are 2 or 4 spaces, and no tabs.  I've used 4 spaces throughout, since I find this easier to read and is similar to Python.  But more people recommend using 2 spaces for both HTML and CSS.  The most common naming style seems to be all lowercase and a hyphen between words.  Spaces are important.  There should be only one space between a css name and its opening bracket "name {".  Also closing bracket needs to line up with the correct section.  More to follow.  -- Jennifer May 4, 2018 --